+++ package.json
@@ -20,9 +20,9 @@
     "googleapis": "^118.0.0",
     "mongodb": "^5.1.0",
     "mongoose": "^7.1.0",
-    "react": "18.2.0",
+    "react": "^18.2.0",
     "react-dom": "18.2.0",
-    "react-native": "0.71.4",
+    "react-native": "^0.71.7",
     "react-native-gesture-handler": "^2.9.0",
     "react-native-reanimated": "~2.14.4",
     "react-native-safe-area-context": "^4.5.0",
