+++ src/screens/Colony Search/ColonySearch.js
@@ -1,0 +1,21 @@
+import { FlatList, Text, View, TouchableHighlight, Image } from "react-native";
+import styles from "./styles";
+import React, { useState } from 'react';
+import { ImageBackground } from 'react-native';
+import { StatusBar } from 'react-native';
+
+
+
+export default function ColonySearch() {
+  return (
+    <ImageBackground
+      source={require("../../../assets/beesbackground.jpg")}
+      style={styles.background}
+    >
+      <View style={styles.container}>
+        <Text style={styles.title}>colonysearch</Text>
+        <StatusBar style="auto" />
+      </View>
+    </ImageBackground>
+  );
+}
